requirements.txt
requests
python-dotenv
supabase
openpyxl